\name{FuseBox-package}
\alias{FuseBox-package}
\alias{FuseBox}
\docType{package}
\title{
Combining classifiers into a single ensemble model
}
\description{
\code{fuse} provides a platform to take existing predictive models and combine there predictions into a single outcome.

}
\details{
asd
}
\author{
Max Kuhn: <max.kuhn@pfizer.com>
}
\references{
refs

}
\keyword{ package }
