\name{predict.fuse}
\alias{predict.fuse}
\title{
Predictions from a combined classifer model
}
\description{
This funciton is used to predict objects created with the \code{fuse} function.
}
\usage{
\method{predict}{fuse}(object, newdata, ...)
}

\arguments{
  \item{object}{
an object of class "fuse"
}
  \item{newdata}{
a data frame of samples for prediction
}
  \item{\dots}{
not currently used
}
}
\details{
details!

}
\value{
a factor vector 
}

\author{
Max Kuhn
}

\seealso{
\code{\link{fuse}}
}
\examples{
2+2
}
\keyword{ model }

